Consolidate duplicate FORM_REGISTRY into a shared formRegistry module
FormPage and FormPreview each maintained their own registry mapping form IDs
to metadata/components. Extract these into a single formRegistry.js that holds
all per-form data (title, labels, description, downloadFilename, component),
eliminating the duplication and making it a single place to register new forms.
