+++ client/src/forms/FormPage.jsx
@@ -5,29 +5,7 @@ import { IconArrowLeft, IconDownload, IconFileTypePdf, IconInfoCircle } from '@t
 import { useQuery } from '@tanstack/react-query';
 
 import Api from '@/Api';
-
-// Registry mapping form IDs (from the URL) to form-specific metadata.
-// Add new form entries here as they are created.
-const FORM_REGISTRY = {
-  cert: {
-    title: 'Certificate of Release',
-    generateLabel: 'Generate Certificate of Release',
-    description: (name) => `SF Sheriff's Dept Certificate of Release for ${name}`,
-    downloadFilename: (deflectionId) => `cert-Certificate-of-Release-${deflectionId}.pdf`,
-  },
-  '849b': {
-    title: 'SFSO 849(b) Report',
-    generateLabel: 'Generate SFSO 849(b) Report',
-    description: (name) => `SFSO 849(b) Report for ${name}`,
-    downloadFilename: (deflectionId) => `849b-report-${deflectionId}.pdf`,
-  },
-  '647f': {
-    title: 'SFSO 647(f) Report',
-    generateLabel: 'Generate SFSO 647(f) Report',
-    description: (name) => `SFSO 647(f) Report for ${name}`,
-    downloadFilename: (deflectionId) => `647f-report-${deflectionId}.pdf`,
-  }
-};
+import FORM_REGISTRY from './formRegistry';
 
 const isReleased = (deflection) => !!deflection?.releasedAt;
